Update and rename README.md to Test Engineer-Hengelo
Job Description

About the job

As a Test Engineer you contribute to a multidisciplinary and self-organizing Agile/Scrum team. The teams are fully responsible for various aspects, such as requirements analysis, architecture, design, software development, testing, validation and product demonstrations. The use of high-tech techniques and test methods is important. The software development process is supported by a modern development environment with full integrated test tooling.

The main tasks and responsibilities for a Test Engineer are:

Review of system requirements to ensure testability;
Contribute to define the test strategy;
Design and implement manual and automated test procedures;
Contribute to improve the test automation framework;
Execute test procedures and report test results;
Client offers careers, not just jobs. You will have the opportunity to develop towards senior expert in your area of expertise. Or if you want to broaden your scope and increase domain knowledge, you could become one of our Software Architects. Or if you are an excellent organizer, you may find that a career path in the field of Agile ScrumMaster fits your profile.

About you

We are searching for a colleague who:

Has knowledge of software development processes;
Has a drive to deliver high quality products;
Is creative, open, flexible;
Is a team-player;
Wants to contribute to a safe and secure society;
Graduated at Bachelor or Master level in the area of software engineering, information/computer science or similar with a talent for Java;
Is capable of rapidly mastering new domains and techniques;
Dutch nationality.
What's in it for you

Become part of a great team in a high tech, innovative international organization
(International) opportunities for growth in tech-, management- and commercial roles
Attention for personal development and own initiative
A nice workplace in a modern building, where we value a good work-life balance and flexibility
40 leave days per year, travel expenses and a flexible benefits à la carte package with options like purchase/sale of leave hours, internet costs, bicycle plan, company fitness, trade union contribution or relocation costs
Interested?

.